grad-milestone-second/src/main/java/hu/geralt/services/beer/beer/BeerCsvServiceImpl.java
package hu.geralt.services.beer.beer;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.FileReader;
import java.util.List;

import com.opencsv.bean.CsvToBeanBuilder;
import hu.geralt.entities.beer.BeerCsvRecord;
import org.springframework.stereotype.Service;

@Service
public class BeerCsvServiceImpl implements BeerCsvService {

    @Override
    public List<BeerCsvRecord> convertCsv(File csvFile) {
        try {
            return new CsvToBeanBuilder<BeerCsvRecord>(new FileReader(csvFile))
                    .withType(BeerCsvRecord.class)
                    .build().parse();
        } catch (FileNotFoundException e) {
            throw new RuntimeException(e);
        }
    }

}
